Projects/50.AI-resume-analyzer/requirements.txt
streamlit
requests
pdfplumber
python-docx
sentence-transformers
faiss-cpu
google-generativeai
Flask
Flask-Cors
awsgi # Bộ chuyển đổi WSGI cho Lambda
gunicorn # Để chạy Flask cục bộ một cách ổn định (tùy chọn)